requirements.txt
asgiref==3.5.2
Django==4.1.1
django-phonenumber-field==7.0.0
phonenumbers==8.12.55
Pillow==9.2.0
python-decouple==3.6
sqlparse==0.4.2
